Move DEBUG to env
