+++ src/Main.ts
@@ -5,7 +5,6 @@ import BoardScene from "./scenes/BoardScene";
 import SplashScene from "./scenes/SplashScene";
 import ResultScene from "./scenes/ResultScene";
 
-export const DEBUG = false;
 export const BAD_COLOR = 0xff0000;
 export const GOOD_COLOR = 0x00ff00;
