+++ src/env.ts
@@ -1,1 +1,2 @@
-export const SERVER_URL = "https://gamejay.cap.dabe.tech";
+export const SERVER_URL = "http://localhost:3000";
+export const DEBUG = false;
